commit-message-generator/requirements.txt
openai>=1.0.0
GitPython>=3.1.30
python-dotenv>=1.0.0
